+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.4
+tokeninfo, return pickup
+
+## 0.1.3
+added authorization_code grant, more fields
+
 ## 0 1.2
 added some fields